misc: Move providers
